+++ lapce-ui/src/lib.rs
@@ -20,6 +20,7 @@ pub mod problem;
 pub mod scroll;
 pub mod search;
 pub mod settings;
+pub mod signature;
 pub mod source_control;
 pub mod split;
 pub mod status;